Check for custom potion effects before falling back on damage value. Fixes BUKKIT-3635
When getting the collection of PotionEffects from a CraftThrownPotion, the
previous implementation only used the damage value to determine the effects,
thus ignoring any custom effects. The implementation was changed to instead
first check for any custom potion effects to override those applied by damage
value, as this is what vanilla Minecraft does when applying the effects.